+++ open_alchemy/column_factory/array_ref/foreign_key.py
@@ -13,6 +13,7 @@ from .. import object_ref
 def set_(
     *,
     ref_model_name: str,
+    logical_name: str,
     model_schema: types.Schema,
     schemas: types.Schemas,
     fk_column: str,
@@ -34,6 +35,7 @@ def set_(
 
     Args:
         ref_model_name: The name of the referenced model.
+        logical_name: The logical name of the array reference property.
         model_schema: The schema which contains the one to many relationship.
         schemas: All the model schemas used to look for the referenced model and to
             resolve any $ref.
@@ -51,8 +53,13 @@ def set_(
     ref_schema = helpers.all_of.merge(schema=ref_schema, schemas=schemas)
 
     # Calculate foreign key artifacts
+    tablename = helpers.peek.tablename(schema=model_schema, schemas={})
+    fk_input_logical_name = f"{tablename}_{logical_name}"
     fk_logical_name, fk_artifacts = object_ref.foreign_key.gather_artifacts(
-        model_schema=model_schema, schemas=schemas, fk_column=fk_column
+        model_schema=model_schema,
+        logical_name=fk_input_logical_name,
+        schemas=schemas,
+        fk_column=fk_column,
     )
 
     # Check whether the foreign key has already been defined in the referenced model