+++ tests/examples/test_example_specs.py
@@ -241,7 +241,7 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
 @pytest.mark.parametrize(
     "filename, model_names, attrs",
     [
-        (
+        pytest.param(
             "relationship/many_to_one/example-spec.yml",
             ("Employee", "Division"),
             {
@@ -249,8 +249,9 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
                 "name": "employee 1",
                 "division": {"id": 12, "name": "division 1"},
             },
+            id="relationship-many-to-one",
         ),
-        (
+        pytest.param(
             "relationship/many_to_one/backref-example-spec.yml",
             ("Employee", "Division"),
             {
@@ -258,8 +259,9 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
                 "name": "employee 1",
                 "division": {"id": 12, "name": "division 1"},
             },
+            id="relationship-many-to-one-backref",
         ),
-        (
+        pytest.param(
             "relationship/many_to_one/custom-foreign-key-example-spec.yml",
             ("Employee", "Division"),
             {
@@ -267,8 +269,9 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
                 "name": "employee 1",
                 "division": {"id": 12, "name": "division 1"},
             },
+            id="relationship-many-to-one-custom-foreign-key",
         ),
-        (
+        pytest.param(
             "relationship/many_to_one/not-nullable-example-spec.yml",
             ("Employee", "Division"),
             {
@@ -276,8 +279,9 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
                 "name": "employee 1",
                 "division": {"id": 12, "name": "division 1"},
             },
+            id="relationship-many-to-one-not-nullable",
         ),
-        (
+        pytest.param(
             "relationship/one_to_many/example-spec.yml",
             ("Division", "Employee"),
             {
@@ -285,8 +289,9 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
                 "name": "division 1",
                 "employees": [{"id": 12, "name": "employee 1"}],
             },
+            id="relationship-one-to-many",
         ),
-        (
+        pytest.param(
             "relationship/one_to_one/example-spec.yml",
             ("Employee", "PayInfo"),
             {
@@ -294,8 +299,9 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
                 "name": "employee 1",
                 "pay_info": {"id": 12, "account_number": "account 1"},
             },
+            id="relationship-one-to-one",
         ),
-        (
+        pytest.param(
             "relationship/many_to_many/example-spec.yml",
             ("Employee", "Project"),
             {
@@ -303,16 +309,8 @@ def test_table_args(engine, filename, model_name, sql, expected_contents):
                 "name": "employee 1",
                 "projects": [{"id": 12, "name": "project 1"}],
             },
+            id="relationship-many-to-many",
         ),
-    ],
-    ids=[
-        "relationship-many-to-one",
-        "relationship-many-to-one-backref",
-        "relationship-many-to-one-custom-foreign-key",
-        "relationship-many-to-one-not-nullable",
-        "relationship-one-to-many",
-        "relationship-one-to-one",
-        "relationship-many-to-many",
     ],
 )
 @pytest.mark.example