+++ tests/open_alchemy/column_factory/array_ref/test_foreign_key.py
@@ -23,6 +23,7 @@ def test_schemas_missing():
     with pytest.raises(exceptions.MalformedRelationshipError):
         array_ref._foreign_key.set_(
             ref_model_name="RefSchema",
+            logical_name="logical name 1",
             model_schema=model_schema,
             schemas={},
             fk_column=fk_column,
@@ -57,6 +58,7 @@ def test_schemas(schemas):
 
     array_ref._foreign_key.set_(
         ref_model_name=ref_model_name,
+        logical_name="logical name 1",
         model_schema=model_schema,
         schemas=schemas,
         fk_column=fk_column,
@@ -69,7 +71,7 @@ def test_schemas(schemas):
                 {
                     "type": "object",
                     "properties": {
-                        f"{tablename}_{fk_column}": {
+                        f"{tablename}_logical name 1_{fk_column}": {
                             "type": "integer",
                             "x-foreign-key": f"{tablename}.{fk_column}",
                             "x-dict-ignore": True,
@@ -104,12 +106,13 @@ def test_models(mocked_facades_models: mock.MagicMock):
 
     array_ref._foreign_key.set_(
         ref_model_name=ref_model_name,
+        logical_name="logical name 1",
         model_schema=model_schema,
         schemas=schemas,
         fk_column=fk_column,
     )
 
-    added_fk_column = getattr(mock_ref_model, f"{tablename}_{fk_column}")
+    added_fk_column = getattr(mock_ref_model, f"{tablename}_logical name 1_{fk_column}")
     assert isinstance(added_fk_column.type, facades.sqlalchemy.column.Integer)
     foreign_key = list(added_fk_column.foreign_keys)[0]
     assert f"{tablename}.{fk_column}" in str(foreign_key)
@@ -137,13 +140,17 @@ def test_fk_def():
             "type": "object",
             "x-tablename": "ref_schema",
             "properties": {
-                "schema_id": {"type": "integer", "x-foreign-key": "schema.id"}
+                f"{tablename}_logical name 1_id": {
+                    "type": "integer",
+                    "x-foreign-key": "schema.id",
+                }
             },
         }
     }
 
     array_ref._foreign_key.set_(
         ref_model_name=ref_model_name,
+        logical_name="logical name 1",
         model_schema=model_schema,
         schemas=schemas,
         fk_column=fk_column,
@@ -154,7 +161,10 @@ def test_fk_def():
             "type": "object",
             "x-tablename": "ref_schema",
             "properties": {
-                "schema_id": {"type": "integer", "x-foreign-key": "schema.id"}
+                f"{tablename}_logical name 1_id": {
+                    "type": "integer",
+                    "x-foreign-key": "schema.id",
+                }
             },
         }
     }