+++ tests/open_alchemy/column_factory/array_ref/test_link.py
@@ -20,6 +20,7 @@ def test_schemas():
     ref_schema = {"type": "object", "x-tablename": "ref_schema", "properties": {}}
     artifacts = types.ObjectArtifacts(
         spec=copy.deepcopy(ref_schema),
+        logical_name="ref_schema",
         fk_column="id",
         relationship=types.RelationshipArtifacts(model_name="RefSchema"),
     )
@@ -42,7 +43,7 @@ def test_schemas():
                 {
                     "type": "object",
                     "properties": {
-                        f"{tablename}_id": {
+                        f"{tablename}_ref_schema_id": {
                             "type": "integer",
                             "x-foreign-key": f"{tablename}.id",
                             "x-dict-ignore": True,
@@ -69,6 +70,7 @@ def test_secondary(mocked_facades_models):
     secondary = "association"
     artifacts = types.ObjectArtifacts(
         spec=copy.deepcopy(ref_schema),
+        logical_name="logical name 1",
         fk_column="id",
         relationship=types.RelationshipArtifacts(
             model_name="RefSchema", secondary=secondary