+++ tests/open_alchemy/column_factory/array_ref/test_schema.py
@@ -13,6 +13,7 @@ from open_alchemy.column_factory import array_ref
         pytest.param(
             types.ObjectArtifacts(
                 spec={},
+                logical_name="logical name 1",
                 fk_column="fk_column",
                 relationship=types.RelationshipArtifacts("RefSchema"),
             ),
@@ -22,6 +23,7 @@ from open_alchemy.column_factory import array_ref
         pytest.param(
             types.ObjectArtifacts(
                 spec={},
+                logical_name="logical name 1",
                 fk_column="fk_column",
                 relationship=types.RelationshipArtifacts("RefSchema"),
                 description="description 1",