+++ tests/open_alchemy/column_factory/object_ref/test_schema.py
@@ -13,6 +13,7 @@ from open_alchemy.column_factory import object_ref
         pytest.param(
             types.ObjectArtifacts(
                 spec={},
+                logical_name="logical name 1",
                 fk_column="fk_column",
                 relationship=types.RelationshipArtifacts("RefSchema"),
             ),
@@ -22,6 +23,7 @@ from open_alchemy.column_factory import object_ref
         pytest.param(
             types.ObjectArtifacts(
                 spec={},
+                logical_name="logical name 1",
                 fk_column="fk_column",
                 relationship=types.RelationshipArtifacts("RefSchema"),
                 nullable=True,
@@ -32,6 +34,7 @@ from open_alchemy.column_factory import object_ref
         pytest.param(
             types.ObjectArtifacts(
                 spec={},
+                logical_name="logical name 1",
                 fk_column="fk_column",
                 relationship=types.RelationshipArtifacts("RefSchema"),
                 nullable=False,
@@ -42,6 +45,7 @@ from open_alchemy.column_factory import object_ref
         pytest.param(
             types.ObjectArtifacts(
                 spec={},
+                logical_name="logical name 1",
                 fk_column="fk_column",
                 relationship=types.RelationshipArtifacts("RefSchema"),
                 description="description 1",